api/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0